Move the notification API to public namespace
