+++ lib/private/notification/action.php
@@ -22,6 +22,8 @@
 namespace OC\Notification;
 
 
+use OCP\Notification\IAction;
+
 class Action implements IAction {
 
 	/** @var string */
