+++ lib/private/notification/notification.php
@@ -22,6 +22,9 @@
 namespace OC\Notification;
 
 
+use OCP\Notification\IAction;
+use OCP\Notification\INotification;
+
 class Notification implements INotification {
 	/** @var string */
 	protected $app;
