+++ lib/public/iservercontainer.php
@@ -468,7 +468,7 @@ interface IServerContainer {
 	/**
 	 * Get the Notification Manager
 	 *
-	 * @return \OC\Notification\IManager
+	 * @return \OCP\Notification\IManager
 	 * @since 8.2.0
 	 */
 	public function getNotificationManager();
