+++ lib/public/notification/imanager.php
@@ -19,17 +19,13 @@
  *
  */
 
-namespace OC\Notification;
+namespace OCP\Notification;
 
 /**
  * Interface IManager
  *
- * @package OC\Notification
+ * @package OCP\Notification
  * @since 8.2.0
- *
- * DEVELOPER NOTE:
- * The notification api is experimental only in 8.2.0! Do not start using it,
- * if you can not prepare an update for the next version afterwards.
  */
 interface IManager extends IApp, INotifier {
 	/**
