+++ tests/lib/notification/notificationtest.php
@@ -23,7 +23,7 @@ namespace Test\Notification;
 
 
 use OC\Notification\Notification;
-use OC\Notification\INotification;
+use OCP\Notification\INotification;
 use Test\TestCase;
 
 class NotificationTest extends TestCase {
@@ -388,12 +388,12 @@ class NotificationTest extends TestCase {
 
 	public function testCreateAction() {
 		$action = $this->notification->createAction();
-		$this->assertInstanceOf('OC\Notification\IAction', $action);
+		$this->assertInstanceOf('OCP\Notification\IAction', $action);
 	}
 
 	public function testAddAction() {
-		/** @var \OC\Notification\IAction|\PHPUnit_Framework_MockObject_MockObject $action */
-		$action = $this->getMockBuilder('OC\Notification\IAction')
+		/** @var \OCP\Notification\IAction|\PHPUnit_Framework_MockObject_MockObject $action */
+		$action = $this->getMockBuilder('OCP\Notification\IAction')
 			->disableOriginalConstructor()
 			->getMock();
 		$action->expects($this->once())
@@ -412,8 +412,8 @@ class NotificationTest extends TestCase {
 	 * @expectedException \InvalidArgumentException
 	 */
 	public function testAddActionInvalid() {
-		/** @var \OC\Notification\IAction|\PHPUnit_Framework_MockObject_MockObject $action */
-		$action = $this->getMockBuilder('OC\Notification\IAction')
+		/** @var \OCP\Notification\IAction|\PHPUnit_Framework_MockObject_MockObject $action */
+		$action = $this->getMockBuilder('OCP\Notification\IAction')
 			->disableOriginalConstructor()
 			->getMock();
 		$action->expects($this->once())
@@ -426,8 +426,8 @@ class NotificationTest extends TestCase {
 	}
 
 	public function testAddActionSecondPrimary() {
-		/** @var \OC\Notification\IAction|\PHPUnit_Framework_MockObject_MockObject $action */
-		$action = $this->getMockBuilder('OC\Notification\IAction')
+		/** @var \OCP\Notification\IAction|\PHPUnit_Framework_MockObject_MockObject $action */
+		$action = $this->getMockBuilder('OCP\Notification\IAction')
 			->disableOriginalConstructor()
 			->getMock();
 		$action->expects($this->exactly(2))
@@ -444,8 +444,8 @@ class NotificationTest extends TestCase {
 	}
 
 	public function testAddParsedAction() {
-		/** @var \OC\Notification\IAction|\PHPUnit_Framework_MockObject_MockObject $action */
-		$action = $this->getMockBuilder('OC\Notification\IAction')
+		/** @var \OCP\Notification\IAction|\PHPUnit_Framework_MockObject_MockObject $action */
+		$action = $this->getMockBuilder('OCP\Notification\IAction')
 			->disableOriginalConstructor()
 			->getMock();
 		$action->expects($this->once())
@@ -464,8 +464,8 @@ class NotificationTest extends TestCase {
 	 * @expectedException \InvalidArgumentException
 	 */
 	public function testAddParsedActionInvalid() {
-		/** @var \OC\Notification\IAction|\PHPUnit_Framework_MockObject_MockObject $action */
-		$action = $this->getMockBuilder('OC\Notification\IAction')
+		/** @var \OCP\Notification\IAction|\PHPUnit_Framework_MockObject_MockObject $action */
+		$action = $this->getMockBuilder('OCP\Notification\IAction')
 			->disableOriginalConstructor()
 			->getMock();
 		$action->expects($this->once())
@@ -478,8 +478,8 @@ class NotificationTest extends TestCase {
 	}
 
 	public function testAddActionSecondParsedPrimary() {
-		/** @var \OC\Notification\IAction|\PHPUnit_Framework_MockObject_MockObject $action */
-		$action = $this->getMockBuilder('OC\Notification\IAction')
+		/** @var \OCP\Notification\IAction|\PHPUnit_Framework_MockObject_MockObject $action */
+		$action = $this->getMockBuilder('OCP\Notification\IAction')
 			->disableOriginalConstructor()
 			->getMock();
 		$action->expects($this->exactly(2))
@@ -512,7 +512,7 @@ class NotificationTest extends TestCase {
 	 * @param bool $expected
 	 */
 	public function testIsValid($isValidCommon, $subject, $expected) {
-		/** @var \OC\Notification\INotification|\PHPUnit_Framework_MockObject_MockObject $notification */
+		/** @var \OCP\Notification\INotification|\PHPUnit_Framework_MockObject_MockObject $notification */
 		$notification = $this->getMockBuilder('\OC\Notification\Notification')
 			->setMethods([
 				'isValidCommon',
@@ -543,7 +543,7 @@ class NotificationTest extends TestCase {
 	 * @param bool $expected
 	 */
 	public function testIsParsedValid($isValidCommon, $subject, $expected) {
-		/** @var \OC\Notification\INotification|\PHPUnit_Framework_MockObject_MockObject $notification */
+		/** @var \OCP\Notification\INotification|\PHPUnit_Framework_MockObject_MockObject $notification */
 		$notification = $this->getMockBuilder('\OC\Notification\Notification')
 			->setMethods([
 				'isValidCommon',
@@ -588,7 +588,7 @@ class NotificationTest extends TestCase {
 	 * @param bool $expected
 	 */
 	public function testIsValidCommon($app, $user, $timestamp, $objectType, $objectId, $expected) {
-		/** @var \OC\Notification\INotification|\PHPUnit_Framework_MockObject_MockObject $notification */
+		/** @var \OCP\Notification\INotification|\PHPUnit_Framework_MockObject_MockObject $notification */
 		$notification = $this->getMockBuilder('\OC\Notification\Notification')
 			->setMethods([
 				'getApp',
