+++ tests/lib/server.php
@@ -116,7 +116,7 @@ class Server extends \Test\TestCase {
 			['NavigationManager', '\OC\NavigationManager'],
 			['NavigationManager', '\OCP\INavigationManager'],
 			['NotificationManager', '\OC\Notification\Manager'],
-			['NotificationManager', '\OC\Notification\IManager'],
+			['NotificationManager', '\OCP\Notification\IManager'],
 			['UserCache', '\OC\Cache\File'],
 			['UserCache', '\OCP\ICache'],
